requirements.txt
datasets>=2.19
pillow>=10.0
torch>=2.2
torchvision>=0.17
tqdm>=4.66
